chore(module-api): combine both as single entity
